refactor: give ActiveSequencesMultiWorker a passport to kv-router
